+++ interfaces/AW-App/src/app/validation-components/download-data/download-data.component.ts
@@ -5,41 +5,92 @@ import { ValidationComponents } from '../validation-components.enum';
 import { ProgramsServiceApiService } from 'src/app/services/programs-service-api.service';
 import { Storage } from '@ionic/storage';
 
+class ValidationAnswer {
+  id: number;
+  did: string;
+  programId: number;
+  attributeId: number;
+  attribute: string;
+  answer: string | number;
+}
+
 @Component({
   selector: 'app-download-data',
   templateUrl: './download-data.component.html',
   styleUrls: ['./download-data.component.scss'],
 })
 export class DownloadDataComponent implements ValidationComponent {
-
   public downloadReady = false;
+  public downloadAborted = false;
   public nrDownloaded: number;
+
+  public validationData: ValidationAnswer[];
 
   constructor(
     public programsService: ProgramsServiceApiService,
     public conversationService: ConversationService,
-    private storage: Storage,
-  ) { }
+    private storage: Storage
+  ) {}
 
   async ngOnInit() {
     await this.downloadData();
   }
 
   public async downloadData() {
-    const validationData = await this.programsService.downloadData();
-    this.storage.set('validationData', validationData);
-    this.nrDownloaded = this.countUniqueDids(validationData);
+    await this.programsService.downloadData().then(
+      (response) => {
+        this.validationData = response;
+      },
+      () => {
+        this.downloadAborted = true;
+        this.complete();
+      }
+    );
+    await this.storage.set('validationData', this.validationData);
+
+    const myPrograms = await this.getProgramData(this.validationData);
+    await this.storage.set('myPrograms', myPrograms);
+
+    this.nrDownloaded = this.countUniqueDids(this.validationData);
     this.downloadReady = true;
     this.complete();
   }
 
-  public countUniqueDids(array) {
-    const dids = [];
-    for (const record of array) {
-      if (!dids.includes(record.did)) {
-        dids.push(record.did);
+  private async getProgramData(validationData: ValidationAnswer[]) {
+    const programIds = this.getUniqueProgramIds(validationData);
+    const programRequests = [];
+    const myPrograms = [];
+
+    programIds.forEach(async (programId) => {
+      programRequests.push(
+        this.programsService
+          .getProgramById(programId)
+          .then((programData) => myPrograms.push(programData))
+      );
+    });
+    await Promise.all(programRequests);
+
+    return myPrograms;
+  }
+
+  private getUniqueProgramIds(validationData: ValidationAnswer[]) {
+    const programIds = [];
+    validationData.forEach((item) => {
+      if (!programIds.includes(item.programId)) {
+        programIds.push(item.programId);
       }
-    }
+    });
+
+    return programIds;
+  }
+
+  public countUniqueDids(validationData: ValidationAnswer[]): number {
+    const dids = [];
+    validationData.forEach((item) => {
+      if (!dids.includes(item.did)) {
+        dids.push(item.did);
+      }
+    });
     return dids.length;
   }
 
@@ -49,10 +100,9 @@ export class DownloadDataComponent implements ValidationComponent {
 
   complete() {
     this.conversationService.onSectionCompleted({
-      name: ValidationComponents.viewAppointments,
+      name: ValidationComponents.downloadData,
       data: {},
       next: this.getNextSection(),
     });
   }
-
 }