+++ services/121-service/src/scripts/seed-dev.ts
@@ -63,7 +63,6 @@ export class SeedDev implements InterfaceScript {
     credential2.answer = '0';
     await credentialAttributesRepository.save(credential2);
 
-    await this.seedPublish.run();
     await this.connection.close();
   }
 }