+++ tools/webhook.js
@@ -3,12 +3,6 @@ const crypto = require("crypto");
 const child_process = require("child_process");
 const exec = child_process.exec;
 
-const secrets = require("./secrets");
-
-const secret = secrets.secret;
-const repo = process.env.GLOBAL_121_REPO;
-
-
 // ----------------------------------------------------------------------------
 //   Functions/Methods/etc:
 // ----------------------------------------------------------------------------
@@ -19,7 +13,9 @@ const repo = process.env.GLOBAL_121_REPO;
  */
 function deploy(target) {
   exec(
-    `cd ${repo} && sudo ./tools/deploy.sh` + (target) ? ` ${target}` : ``,
+    target
+      ? `cd ${process.env.GLOBAL_121_REPO} && sudo ./tools/deploy.sh "${target}"`
+      : `cd ${process.env.GLOBAL_121_REPO} && sudo ./tools/deploy.sh`,
     function (error, stdout, stderr) {
       if (error) {
         console.log(stderr);
@@ -39,14 +35,13 @@ http
     let body = [];
     req.on("data", function(chunk) {
       body.push(chunk);
-      console.log("BODY: ", body);
     });
     req.on("end", function() {
       let str = Buffer.concat(body).toString();
       let sig =
         "sha1=" +
         crypto
-          .createHmac("sha1", secret)
+          .createHmac("sha1", process.env.GITHUB_WEBHOOK_SECRET)
           .update(str)
           .digest("hex");
       let payload = JSON.parse(str);
@@ -75,6 +70,6 @@ http
     });
     res.end();
   })
-  .listen(3099);
+  .listen(process.env.NODE_PORT);
 
-console.log("Listening on port 3099");
+console.log(`Listening on port ${process.env.NODE_PORT}`);