+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- 2020-04-23: Remove `tools/secrets.json`, set `GITHUB_WEBHOOK_SECRET` in `webhook.service`
+
+
+## [0.2.1] - 2020-04-22
+Test hotfix-release to test automatic deployment of patch-releases.
 
 ## [0.2.0] - 2020-04-22
 
@@ -29,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - 2020-03-11: Remove usernames from `services/121-service/src/secrets.ts` AB#1641
 
-[Unreleased]: https://github.com/global-121/121-platform/compare/v0.2.0...master
+[Unreleased]: https://github.com/global-121/121-platform/compare/v0.2.1...master
+[0.2.1]: https://github.com/global-121/121-platform/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/global-121/121-platform/releases/tag/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/global-121/121-platform/releases/tag/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/global-121/121-platform/releases/tag/v0.1.0